+++ include/OpenImageDenoise/oidn.hpp
@@ -215,12 +215,24 @@ OIDN_NAMESPACE_BEGIN
                                bytePixelStride, byteRowStride);
     }
 
+    // Removes an image parameter of the filter that was previously set.
+    void removeImage(const char* name)
+    {
+      oidnRemoveFilterImage(handle, name);
+    }
+
     // Sets an opaque data parameter of the filter (owned by the user).
     void setData(const char* name,
                  void* ptr, size_t byteSize)
     {
       oidnSetSharedFilterData(handle, name,
                               ptr, byteSize);
+    }
+
+    // Removes an opaque data parameter of the filter that was previously set.
+    void removeData(const char* name)
+    {
+      oidnRemoveFilterData(handle, name);
     }
 
     // Sets a boolean parameter of the filter.